requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
black==24.8.0
click==8.1.8
Django==4.2.19
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
djangorestframework==3.15.2
flake8==7.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
phonenumbers==8.13.55
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.12.2
